+++ Scripts/UnityCore/Components/KiCanvasGroup.cs
@@ -8,6 +8,11 @@ namespace KimicuUtility
     {
         private CanvasGroup _canvasGroup;
 
+        private void Awake()
+        {
+            _canvasGroup ??= GetComponent<CanvasGroup>();
+        }
+
         private void OnValidate()
         {
             _canvasGroup ??= GetComponent<CanvasGroup>();