+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.0.19] - 07.10.2023
+### Add
+- Random queue - list.RandomQueue(amountElements).
+
+### Fixed
+- KiCanvasGroup Component. (Null reference exception!)
+
+
 ## [1.0.18] - 06.10.2023
 ### Fixed
 - Fix RecurringDelay. The delay did not work correctly.